backend/requirements.txt
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.9.0
cachelib==0.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==46.0.3
cycler==0.12.1
Deprecated==1.2.18
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.1.0
eventlet==0.40.3
Flask==3.0.0
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-Limiter==4.0.0
flask-restx==1.3.2
Flask-SocketIO==5.3.6
fonttools==4.60.1
future==1.0.0
gotrue==2.9.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.20.2
mdurl==0.1.2
numpy==2.3.4
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==5.17.0
pluggy==1.6.0
postgrest==0.13.2
psutil==5.9.6
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.2.5
pypng==0.20220715.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.10.0
pytz==2025.2
qrcode==7.4.2
realtime==1.0.6
redis==5.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.31.0
rich==14.2.0
rpds-py==0.28.0
scipy==1.16.2
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
storage3==0.7.7
StrEnum==0.4.15
stripe==7.8.0
supabase==2.3.0
supafunc==0.3.3
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.1.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
